fix: extract shared color-temperature code into `drm_colortemp_utils.{h,c}` and fix a `basename()`/`dirname()` bug
Extract shared color-temperature code into `drm_colortemp_utils.{h,c}` and fix a `basename()`/`dirname()` memory safety issue in the daemon.

- `temp_to_rgb()` and the gamma LUT fill loop were duplicated identically across `drm_colortemp.c` and `drm_colortemp_daemon_inotify.c`
- The daemon used a single `strdup()` for both `dirname()` and `basename()`, which is unsafe since both may modify their argument; additionally the buffer was freed before `filename` was done being used inside the event loop

- New `drm_colortemp_utils.h` / `drm_colortemp_utils.c` with `temp_to_rgb()` and `fill_gamma_luts()`
- `drm_colortemp.c`: removed `temp_to_rgb()`, replaced inline LUT fill with `fill_gamma_luts()` call
- `drm_colortemp_daemon_inotify.c`: same deduplication; fixed `basename()`/`dirname()` to use two separate `strdup()` calls, moved `free()` to after the while-loop
- `Makefile`: added `drm_colortemp_utils.o` to both `TOOL_OBJ` and `DAEMON_OBJ`, added header dependencies

Both binaries build cleanly and behave identically. The shared color-temperature logic lives in one place, and the daemon no longer has a use-after-free risk on the config filename.

--
